+++ plugins/operators/load_fact.py
@@ -13,17 +13,20 @@ class LoadFactOperator(BaseOperator):
                  redshift_conn_id="",
                  table="",
                  sql="",
+                 append=True,
                  *args, **kwargs):
 
         super(LoadFactOperator, self).__init__(*args, **kwargs)
         self.redshift_conn_id = redshift_conn_id
         self.table = table
         self.sql = sql
+        self.append=append
         
 
     def execute(self, context):
         redshift = PostgresHook(postgres_conn_id=self.redshift_conn_id)
-        self.log.info("Clearing data from destination Redshift table")
-        redshift.run("DELETE FROM {}".format(self.table))
+        if self.append == False:
+            self.log.info("Clearing data from destination Redshift table")
+            redshift.run("DELETE FROM {}".format(self.table))
         self.log.info("insert data to destination Redshift table")
         redshift.run(self.sql)
